requirements.txt
fastapi
uvicorn
gunicorn
jsonschema
torch
transformers
pillow
pypdf